treewide: fixed DB and added restart alert back
