+++ eduu/plugins/admins/pins.py
@@ -3,7 +3,8 @@ from pyrogram.types import Message
 
 from eduu.config import PREFIXES
 from eduu.database.admins import check_if_antichannelpin, toggle_antichannelpin
-from eduu.utils import commands, require_admin
+from eduu.utils import commands
+from eduu.utils.decorators import require_admin
 from eduu.utils.localization import use_chat_lang
