+++ eduu/plugins/chats.py
@@ -4,7 +4,7 @@
 from pyrogram import Client
 from pyrogram.types import Message
 
-from eduu.utils import add_chat, chat_exists
+from eduu.database.chats import add_chat, chat_exists
 
 # This is the first plugin run to guarantee
 # that the actual chat is initialized in the DB.
