+++ eduu/plugins/rules.py
@@ -6,7 +6,8 @@ from pyrogram.types import InlineKeyboardMarkup, Message
 
 from eduu.config import PREFIXES
 from eduu.database.rules import get_rules, set_rules
-from eduu.utils import button_parser, commands, require_admin
+from eduu.utils import button_parser, commands
+from eduu.utils.decorators import require_admin
 from eduu.utils.localization import use_chat_lang
