+++ eduu/utils/decorators.py
@@ -1,0 +1,64 @@
+# SPDX-License-Identifier: MIT
+# Copyright (c) 2018-2022 Amano Team
+
+from functools import partial, wraps
+from typing import Union
+
+from pyrogram import Client
+from pyrogram.enums import ChatType
+from pyrogram.types import CallbackQuery, Message
+
+from eduu.utils.localization import (
+    default_language,
+    get_lang,
+    get_locale_string,
+    langdict,
+)
+from eduu.utils.utils import check_perms
+
+
+def require_admin(
+    permissions: Union[list, str] = None,
+    allow_in_private: bool = False,
+    complain_missing_perms: bool = True,
+):
+    def decorator(func):
+        @wraps(func)
+        async def wrapper(
+            client: Client, message: Union[CallbackQuery, Message], *args, **kwargs
+        ):
+            lang = get_lang(message)
+            strings = partial(
+                get_locale_string,
+                langdict[lang].get("admins", langdict[default_language]["admins"]),
+                lang,
+                "admins",
+            )
+
+            if isinstance(message, CallbackQuery):
+                sender = partial(message.answer, show_alert=True)
+                msg = message.message
+            elif isinstance(message, Message):
+                sender = message.reply_text
+                msg = message
+            else:
+                raise NotImplementedError(
+                    f"require_admin can't process updates with the type '{message.__name__}' yet."
+                )
+
+            # We don't actually check private and channel chats.
+            if msg.chat.type == ChatType.PRIVATE:
+                if allow_in_private:
+                    return await func(client, message, *args, *kwargs)
+                return await sender(strings("private_not_allowed"))
+            if msg.chat.type == ChatType.CHANNEL:
+                return await func(client, message, *args, *kwargs)
+            has_perms = await check_perms(
+                message, permissions, complain_missing_perms, strings
+            )
+            if has_perms:
+                return await func(client, message, *args, *kwargs)
+
+        return wrapper
+
+    return decorator
